Update method calls for Wicket 6.
SAK-28995

+++ profile2/tool/src/java/org/sakaiproject/profile2/tool/pages/panels/GalleryImageEdit.java
@@ -64,7 +64,7 @@ public class GalleryImageEdit extends Panel {
 	
 	public GalleryImageEdit(String id,
 			final String userId, final GalleryImage image,
-			final int galleryPageIndex) {
+			final long galleryPageIndex) {
 
 		super(id);
